# API Marshaling: Add generated marshalers for RESTXML protocol