--- a/private/model/api/shape.go
+++ b/private/model/api/shape.go
@@ -78,6 +78,9 @@ type Shape struct {
 
 	OrigShapeName string `json:"-"`
 
+	UsedInMap  bool
+	UsedInList bool
+
 	// Defines if the shape is a placeholder and should not be used directly
 	Placeholder bool
 
@@ -185,10 +188,18 @@ func (s *Shape) GoStructValueType(name string, ref *ShapeRef) string {
 	return v
 }
 
+func (s *Shape) IsRefPayload(name string) bool {
+	return s.Payload == name
+}
+
+func (s *Shape) IsRefPayloadReader(name string, ref *ShapeRef) bool {
+	return (ref.Streaming || ref.Shape.Streaming) && s.IsRefPayload(name)
+}
+
 // GoStructType returns the type of a struct field based on the API
 // model definition.
 func (s *Shape) GoStructType(name string, ref *ShapeRef) string {
-	if (ref.Streaming || ref.Shape.Streaming) && s.Payload == name {
+	if s.IsRefPayloadReader(name, ref) {
 		rtype := "io.ReadSeeker"
 		if strings.HasSuffix(s.ShapeName, "Output") {
 			rtype = "io.ReadCloser"
@@ -513,7 +524,9 @@ func (s *Shape) NestedShape() *Shape {
 }
 
 var structShapeTmpl = template.Must(template.New("StructShape").Funcs(template.FuncMap{
-	"GetCrosslinkURL": GetCrosslinkURL,
+	"GetCrosslinkURL":      GetCrosslinkURL,
+	"MarshalShapeGoCode":   MarshalShapeGoCode,
+	"UnmarshalShapeGoCode": UnmarshalShapeGoCode,
 }).Parse(`
 {{ .Docstring }}
 {{ if ne $.OrigShapeName "" -}}
@@ -597,6 +610,13 @@ func (s *{{ $builderShapeName }}) get{{ $name }}() (v {{ $context.GoStructValueT
 
 {{ end }}
 {{ end }}
+
+{{ if not $.API.NoGenMarshalers -}}
+{{ MarshalShapeGoCode $ }}
+{{- end }}
+{{ if not $.API.NoGenUnmarshalers -}}
+{{ UnmarshalShapeGoCode $ }}
+{{- end }}
 `))
 
 var enumShapeTmpl = template.Must(template.New("EnumShape").Parse(`