… private/protocol/fields.go
… private/protocol/header_encoder.go
… private/protocol/path_replace.go
… private/protocol/path_replace_test.go
… private/protocol/query_encoder.go
… private/protocol/rest/encode.go
… private/protocol/rest/encode_test.go
… private/protocol/restxml/build_test.go
… private/protocol/restxml/encode.go
… private/protocol/restxml/encode_test.go
… private/protocol/restxml/restxml.go
… private/protocol/restxml/unmarshal_test.go
… private/protocol/xml/encode.go
… private/protocol/xml/encode_test.go
… service/cloudfront/api.go
… service/route53/api.go
… service/s3/api.go